# DOCS/API.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SARVAL API",
    "version": "1.0.0",
    "description": "API del planificador de descargas en silos de SARVAL.\n\nToda la API está protegida con JWT bearer salvo `POST /api/auth/login`. El motor de simulación se recalcula **automáticamente** en cualquier endpoint que cambie el plan (subir Excel, crear/editar/borrar viaje, parada, box, lectura de nivel, franja de productividad o parámetros de tolva). El único recálculo manual con notificación de webhooks es `POST /api/planning/recalculate`.\n\n**Concepto \"plan activo\"**: la mayoría de endpoints aceptan `?plan_id=…`. Si no se pasa, se usa el plan vigente (el de la semana en curso).\n\n**Semana operativa**: Lunes 06:00 → Sábado 22:00 (consumo continuo, también de noche).\n\nVer `DOCS/API.md` en el repo para los flujos típicos."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local"
    },
    {
      "url": "https://sarval.example.com",
      "description": "Producción (ajustar)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Stoppage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "plan_id": {
            "type": "integer"
          },
          "tolva_id": {
            "type": "integer"
          },
          "tipo": {
            "type": "string",
            "example": "mantenimiento"
          },
          "descripcion": {
            "type": "string",
            "example": "Cambio de filtros"
          },
          "dia": {
            "type": "string",
            "example": "Martes"
          },
          "hora_inicio": {
            "type": "string",
            "example": "08:00"
          },
          "hora_fin": {
            "type": "string",
            "example": "10:00"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "responses": {
...
      "NotFound": {
        "description": "Recurso no encontrado.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Petición inválida.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/stoppages": {
      "get": {
        "tags": [
          "stoppages"
        ],
        "summary": "Listar paradas del plan",
        "parameters": [
          {
            "name": "plan_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Plan a consultar. Si se omite se usa el plan activo (vigente)."
          },
          {
            "name": "tolva_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Filtra por tolva."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Stoppage"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "stoppages"
        ],
        "summary": "Crear parada (recalcula)",
        "description": "Pausa TODA actividad (consumo, camiones y boxes) entre `hora_inicio` y `hora_fin` del día indicado.",
        "parameters": [
          {
            "name": "plan_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Plan a consultar. Si se omite se usa el plan activo (vigente)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "tolva_id",
                  "dia",
                  "hora_inicio",
                  "hora_fin"
                ],
                "properties": {
                  "tolva_id": {
                    "type": "integer"
                  },
                  "tipo": {
                    "type": "string",
                    "example": "mantenimiento"
                  },
                  "descripcion": {
                    "type": "string"
                  },
                  "dia": {
                    "type": "string",
                    "example": "Martes"
                  },
                  "hora_inicio": {
                    "type": "string",
                    "example": "08:00"
                  },
                  "hora_fin": {
                    "type": "string",
                    "example": "10:00"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stoppage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/stoppages/{id}": {
      "put": {
        "tags": [
          "stoppages"
        ],
        "summary": "Editar parada (recalcula)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tipo": {
                    "type": "string"
                  },
                  "descripcion": {
                    "type": "string"
                  },
                  "dia": {
                    "type": "string"
                  },
                  "hora_inicio": {
                    "type": "string"
                  },
                  "hora_fin": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stoppage"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "stoppages"
        ],
        "summary": "Eliminar parada (recalcula)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Eliminado."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  }
}